ecourseapisv2/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
cryptography==46.0.6
Django==4.2.28
django-ckeditor==6.7.3
django-js-asset==2.2.0
django-oauth-toolkit==3.2.0
djangorestframework==3.16.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==26.0
pycparser==3.0
PyMySQL==1.1.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
